LIBRE !!!!!!!!!!!!!!!

+++ routes/player/index.ts
@@ -2,13 +2,18 @@ import express from 'express';
 import pagesize from "../../queryConfig";
 import { PlayerRepository } from '../../src/interfaces/repository/player.repository';
 import { PrismaPlayerRepository } from '../../src/implementation/repository/prisma/player.repository';
-import { Schemas } from '../../schemas/index'
+import { Schemas } from '../../schemas/index';
 
 const router = express.Router();
 const repository:PlayerRepository = new PrismaPlayerRepository();
 
 // Get all players
 router.get("/", async (req: any, res: any) => {
+    var { error, value } = Schemas.playersGetSchema.validate(req.query);
+    if (error !== undefined) {
+        res.status(300).send("Error: "+ error);
+        return;
+    }
     // Paramètres de requête pour la pagination : page et size, des entiers positifs
     var {page, size} = req.query;
     if(page===undefined || page<=0){
@@ -20,12 +25,22 @@ router.get("/", async (req: any, res: any) => {
         res.status(400).send("Error: "+err);
         return;
     });
-    console.log(Schemas.playerSchema.validate(players));
-    res.status(200).json(players);
+    var { error, value } = Schemas.playersSchema.validate(players);
+    if (error === undefined) {
+        res.status(200).json(players);
+    }
+    else {
+        res.status(500).send("Error: "+ error);
+    }
 });
 
 // Get player by id
 router.get("/:id", async (req: any, res: any) => {
+    var { error, value } = Schemas.playersGetByIdSchema.validate(req.params);
+    if (error !== undefined) {
+        res.status(300).send("Error: "+ error);
+        return;
+    }
     // id: entier positif, id du joueur à renvoyer
     const { id } = req.params;
     const player = await repository.findById(parseInt(id)).catch((err)=>{
@@ -36,11 +51,22 @@ router.get("/:id", async (req: any, res: any) => {
         res.status(404).send("Player not found");
         return;
     }
-    res.status(200).json(player);
+    var { error, value } = Schemas.playerSchema.validate(player);
+    if (error === undefined) {
+        res.status(200).json(player);
+    }
+    else {
+        res.status(500).send("Error: "+ error);
+    }
 });
 
 // Get player by name
 router.get("/name/:name", async (req: any, res: any) => {
+    var { error, value } = Schemas.playersGetByNameSchema.validate(req.params);
+    if (error !== undefined) {
+        res.status(300).send("Error: "+ error);
+        return;
+    }
     // name: string, nom du joueur à renvoyer
     const { name } = req.params;
     const player = await repository.findByName(name).catch((err)=>{
@@ -51,20 +77,42 @@ router.get("/name/:name", async (req: any, res: any) => {
         res.status(404).send("Player not found");
         return;
     }
-    res.status(200).json(player);
+    var { error, value } = Schemas.playerSchema.validate(player);
+    if (error === undefined) {
+        res.status(200).json(player);
+    }
+    else {
+        res.status(500).send("Error: "+ error);
+    }
 });
 
 router.post("/", async (req: any, res: any) => { 
+    var { error, value } = Schemas.playersPostSchema.validate(req.body);
+    if (error !== undefined) {
+        res.status(300).send("Error: "+ error);
+        return;
+    }
     // name: string, nom du joueur à créer
     const { name } = req.body;
     const player = await repository.save(name).catch((err)=>{
         res.status(400).send("Error: "+err);
         return;
     });
-    res.status(201).json(player);
+    var { error, value } = Schemas.playerSchema.validate(player);
+    if (error === undefined) {
+        res.status(201).json(player);
+    }
+    else {
+        res.status(500).send("Error: "+ error);
+    }
 });
 
 router.delete("/:id", async (req: any, res: any)=>{
+    var { error, value } = Schemas.playersDeleteSchema.validate(req.params);
+    if (error !== undefined) {
+        res.status(300).send("Error: "+ error);
+        return;
+    }
     // id: entier positif, id du joueur à supprimer
 	const {id}= req.params;
     const player = await repository.delete(parseInt(id)).catch((err)=>{
@@ -75,8 +123,13 @@ router.delete("/:id", async (req: any, res: any)=>{
         res.status(404).send("Player not found");
         return;
     }
-    res.status(201).json(player);
-	
+    var { error, value } = Schemas.playerSchema.validate(player);
+    if (error === undefined) {
+        res.status(201).json(player);
+    }
+    else {
+        res.status(500).send("Error: "+ error);
+    }
 });
 
 router.put("/:id", async (req: any, res: any)=>{
@@ -84,6 +137,12 @@ router.put("/:id", async (req: any, res: any)=>{
     // name: string, nouveau nom du joueur
 	const {id} = req.params;
     const {name}= req.body;
+
+    var { error, value } = Schemas.playersPutSchema.validate({id, name});
+    if (error !== undefined) {
+        res.status(300).send("Error: "+ error);
+        return;
+    }
     
 	let notValid = !name;
     if(notValid){
@@ -99,7 +158,13 @@ router.put("/:id", async (req: any, res: any)=>{
         res.status(404).send("Player not found or Error while updating player.");
         return;
     }
-    res.status(201).json(player);
+    var { error, value } = Schemas.playerSchema.validate(player);
+    if (error === undefined) {
+        res.status(201).json(player);
+    }
+    else {
+        res.status(500).send("Error: "+ error);
+    }
 	
 });
